Refactoring some things
