--- src/main/java/net/ludocrypt/backrooms/client/BackroomsClient.java
@@ -1,12 +1,0 @@
-package net.ludocrypt.backrooms.client;
-
-import net.fabricmc.api.ClientModInitializer;
-
-public class BackroomsClient implements ClientModInitializer {
-
-	@Override
-	public void onInitializeClient() {
-
-	}
-
-}
