+++ src/main/java/net/synchlabs/backrooms/client/BackroomsClient.java
@@ -1,0 +1,12 @@
+package net.synchlabs.backrooms.client;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class BackroomsClient implements ClientModInitializer {
+
+	@Override
+	public void onInitializeClient() {
+
+	}
+
+}
